Do not use Double.MIN_VALUE as delta for comparing doubles.
`Double.MIN_VALUE` is not suitable as the "delta" value for comparing
two doubles, because it is extremely small and may very well be smaller
than the difference that may be expected due to the imprecise
representation of floating point numbers.

This commit replaces all uses of Double.MIN_VALUE in assertions
comparing doubles by the `DOUBLE_PRECISION` constant defined in the
`BasicTest` class (which is presumably exactly intended for that usage).